Semi-Robust input validation
Implemented semi-robust input validation, it works as follows:

- Program takes user input and stores this in an array of characters using fgets (scanf was used previously but was removed due to safety reasons).
- This input is validated using the is_valid_input() method which ensures the string conforms to the syntax expected from an integer or decimal number.
- if the input matches the expected syntax, the string is converted to a double for use as an operand. If this input is invalid an error message is displayed to the user.

NOTE: Many methods were tried to validate the input, however as this input is used directly as an operand, I could not validate it by using 'atof' or 'strtod' as these return 0.0 floating point values on unsuccessful conversion, which could be a valid input in the context of my program. Moreover, 'fgets' stores values read from the file (standard input) as characters therefore validation must occur before the conversion to a double value.